requirements.txt
# DAIA v5 - Hugging Face Spaces 배포용
# llama-cpp-python / CUDA 제거 → Gemini API 사용

# ── UI ───────────────────────────────────────────────────────
gradio>=4.20.0

# ── 데이터 처리 ──────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.4.0

# ── 시각화 ───────────────────────────────────────────────────
plotly>=5.17.0
kaleido>=0.2.1

# ── 파일 형식 ─────────────────────────────────────────────────
openpyxl>=3.1.0
pyarrow>=14.0.0

# ── 대용량 파일 ───────────────────────────────────────────────
dask[dataframe]>=2024.1.0
duckdb>=0.10.0

# ── 설정 / 유틸 ───────────────────────────────────────────────
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# ── 리포트 ────────────────────────────────────────────────────
reportlab>=4.0.0
pillow>=10.0.0
beautifulsoup4>=4.12.0

# ── LLM: Gemini API ──────────────────────────────────────────
google-generativeai>=0.5.0

# ── 사용량 저장: Supabase ─────────────────────────────────────
supabase>=2.3.0

# ── 로컬 개발용 시크릿 로드 ───────────────────────────────────
python-dotenv>=1.0.0
